Change the type of sample_rate to int
We can assume that int is at least 32 bits long and
it is the optimal variable type on the current arch.
There is no need to use fixed length variable type,
when we dont care about the length and all possible values
of the variable fit to int.

+++ cinelerra/edlsession.h
@@ -164,7 +164,7 @@ public:
 	int64_t record_write_length;
 // Show title and action safe regions in CWindow
 	int safe_regions;
-    int64_t sample_rate;
+	int sample_rate;
 	float scrub_speed;
 // Load files as a duration
 	int si_useduration;
